VEX_Notifier/requirements.txt
#   Locked Versions
openai==1.0.0                       #   OpenAI 
openai-whisper==20230930            #   Audio Transcription
discord.py==2.0.0                   #   Discord Bot
yt-dlp==2024.6.9                    #   YouTube downloads

#   Flexible Versions
google-api-python-client>=2.0.0     #   Youtube API
requests>=2.31.0                    #   HTTP requests
python-dotenv>=1.0.0                #   .env files
tenacity>=8.0.0                     #   Retries failed downloads